// src/api/models/ProgramarMantenimientoRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ProgramarMantenimientoRequest = {
    tipoMantenimiento: string;
    prioridad: string;
    descripcionDelProblema: string;
    condicionGeneral: string;
    tecnicoAsignadoId?: number | null;
    observaciones?: string | null;
    requiereAprobacionCliente?: boolean;
};